requirements.txt
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
boto3==1.40.13
botocore==1.40.13
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
crowdai_api==0.1.22
cycler==0.12.1
decorator==5.2.1
executing==2.2.0
fastenum==1.1.2
flatland-rl==4.1.4
fonttools==4.59.1
idna==3.10
importlib_resources==6.5.2
ipycanvas==0.14.1
ipyevents==2.0.2
ipython==8.37.0
ipywidgets==8.1.7
jedi==0.19.2
jmespath==1.0.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lxml==6.0.0
matplotlib==3.10.5
matplotlib-inline==0.1.7
msgpack==1.1.1
msgpack-numpy==0.4.8
networkx==3.4.2
numpy==1.26.4
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
prompt_toolkit==3.0.51
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
pyglet==2.1.8
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-gitlab==6.2.0
pytz==2025.2
PyYAML==6.0.2
recordtype==1.4
redis==6.4.0
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.13.1
six==1.17.0
stack-data==0.6.3
svgutils==0.3.4
timeout-decorator==0.5.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
